# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/orgs/{slug}/admin/city/status": {
      "get": {
        "operationId": "CityController_status",
        "parameters": [
          {
            "name": "view",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "City"
        ]
      }
    },
...
    "/v1/orgs/{slug}/admin/city/tenant/{orgSlug}": {
      "get": {
        "operationId": "CityController_tenant",
        "parameters": [
          {
            "name": "orgSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "City"
        ]
      }
    },
...
    "/v1/orgs/{slug}/admin/city/mail/{messageId}/reply": {
      "post": {
        "operationId": "CityController_reply",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "City"
        ]
      }
    },
    "/v1/orgs/{slug}/admin/city/mail/{messageId}/read": {
      "post": {
        "operationId": "CityController_markRead",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "City"
        ]
      }
    },
    "/v1/orgs/{slug}/admin/city/mail/{messageId}/mark-unread": {
      "post": {
        "operationId": "CityController_markUnread",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "City"
        ]
      }
    },
    "/v1/orgs/{slug}/admin/city/mail/{messageId}/archive": {
      "post": {
        "operationId": "CityController_archive",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "City"
        ]
      }
    }
  },
  "info": {
    "title": "Mergecrew API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}